refactor: App.tsx에서 AppContent 분리
